pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "urgency"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.2",
    "torchvision>=0.17",
    "timm>=1.0",
    "lightning>=2.2",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "pandas>=2.2",
    "scikit-learn>=1.4",
    "matplotlib>=3.8",
    "Pillow>=10.2",
    "numpy>=1.26",
    "rich>=13.7",
    "typer>=0.12",
    "python-multipart>=0.0.9",
    "scipy>=1.13",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1",
    "ruff>=0.4",
    "pre-commit>=3.7",
]

[project.scripts]
urgency = "urgency.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["urgency"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
